Add test. Rename index.jsx to typography.jsx
